updating models to allow for skate configs, boots, blades
